ep: Move scores aggegation to edit_prediction_metrics
This way it can be shared with Python bindings.
