+++ crates/edit_prediction_cli/src/metrics.rs
@@ -1,7 +1,5 @@
 #![allow(unused_imports)]
 
-use crate::example::ActualCursor;
-
 pub use edit_prediction_metrics::ClassificationMetrics;
 pub use edit_prediction_metrics::Counts;
 pub use edit_prediction_metrics::DeltaChrFMetrics;
@@ -14,11 +12,5 @@ pub use edit_prediction_metrics::delta_chr_f;
 pub use edit_prediction_metrics::delta_chr_f_beta;
 pub use edit_prediction_metrics::exact_lines_match;
 pub use edit_prediction_metrics::extract_changed_lines_from_diff;
+pub use edit_prediction_metrics::has_isolated_whitespace_changes;
 pub use edit_prediction_metrics::is_editable_region_correct;
-
-pub fn has_isolated_whitespace_changes(patch_str: &str, cursor: Option<&ActualCursor>) -> bool {
-    edit_prediction_metrics::has_isolated_whitespace_changes(
-        patch_str,
-        cursor.map(|cursor| cursor.row),
-    )
-}
