+++ crates/edit_prediction_metrics/src/edit_prediction_metrics.rs
@@ -1,6 +1,8 @@
 mod kept_rate;
 mod patch_metrics;
+mod prediction_score;
 mod reversal;
+mod summary;
 mod tokenize;
 mod tree_sitter;
 
@@ -22,5 +24,10 @@ pub use patch_metrics::extract_changed_lines_from_diff;
 pub use patch_metrics::has_isolated_whitespace_changes;
 pub use patch_metrics::is_editable_region_correct;
 pub use patch_metrics::reconstruct_texts_from_diff;
+pub use prediction_score::{
+    ActualPredictionCursor, PredictionReversalContext, PredictionScore, PredictionScoringInput,
+    PrepareExpectedPatchError, PreparedExpectedPatch, prepare_expected_patches, score_prediction,
+};
 pub use reversal::compute_prediction_reversal_ratio_from_history;
+pub use summary::{PredictionSummaryInput, QaSummaryData, SummaryJson, compute_summary};
 pub use tree_sitter::count_tree_sitter_errors;
